mobcoins are now stored as longs instead of ints

+++ src/main/java/me/swanis/mobcoins/events/MobCoinsShopEvent.java
@@ -14,9 +14,9 @@ public class MobCoinsShopEvent extends Event implements Cancellable {
 
     private Profile profile;
     private Reward reward;
-    private int price;
+    private long price;
 
-    public MobCoinsShopEvent(Profile profile, Reward reward, int price) {
+    public MobCoinsShopEvent(Profile profile, Reward reward, long price) {
         this.profile = profile;
         this.reward = reward;
         this.price = price;
@@ -30,11 +30,11 @@ public class MobCoinsShopEvent extends Event implements Cancellable {
         return reward;
     }
 
-    public int getPrice() {
+    public long getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(long price) {
         this.price = price;
     }
